Fixed user_config_t::raw type on my_vial.

+++ keyboards/green_keys/gravity_45/keymaps/my_vial/my_rgb.c
@@ -2,7 +2,7 @@
 #include <my_keycodes.h>
 
 typedef union {
-    uint8_t raw;
+    uint32_t raw;
     struct {
         bool is_rgb_per_layer;
     };
